Update src/chinstrap/views/common.clj
The CDN served up from code.jquery.org was a untrusted/self-sign certificate. Changed to the correct, trusted certificate available off Google's CDN hosted.  

Please note that we have updated to point-releases of jquery and jquery-ui that were not tested previously with the application (previously tested with jquery 1.8.2 and jquery-ui 1.9.0).